Restructure import_module! macro in order to avoid unstable features
