+++ rust/runtime/Cargo.toml
@@ -28,7 +28,7 @@ authors = ["TVM Contributors"]
 edition = "2018"
 
 [dependencies]
-bounded-spsc-queue = "0.4"
+crossbeam = "0.7.3"
 failure = "0.1"
 itertools = "0.8"
 lazy_static = "1.4"
